requirements.txt
pillow
python-dateutil
PyYAML
requests
websocket-client
